+++ include/render/textures/CheckerBoardTexture.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <memory>
 #include "render/textures/Texture.h"
 #include "core/Color.h"
@@ -25,37 +27,45 @@ namespace render {
   class CheckerBoardTexture : public Texturec {
   public:
     inline explicit CheckerBoardTexture(TextureMapping2D* mapping)
-      : m_mapping(mapping),
-        m_brightTexture(nullptr),
-        m_darkTexture(nullptr)
-    {
+        : m_mapping(mapping),
+          m_brightTexture(nullptr),
+          m_darkTexture(nullptr) {
     }
-    
-    inline explicit CheckerBoardTexture(TextureMapping2D* mapping, std::shared_ptr<Texturec> brightTexture, std::shared_ptr<Texturec> darkTexture)
-      : m_mapping(mapping),
-        m_brightTexture(brightTexture),
-        m_darkTexture(darkTexture)
-    {
+
+    inline explicit CheckerBoardTexture(TextureMapping2D* mapping,
+                                        std::shared_ptr<Texturec> brightTexture,
+                                        std::shared_ptr<Texturec> darkTexture)
+        : m_mapping(mapping),
+          m_brightTexture(brightTexture),
+          m_darkTexture(darkTexture) {
     }
-    
+
     inline std::shared_ptr<Texturec> brightTexture() const {
       return m_brightTexture;
     }
-    
+
     inline void setBrightTexture(std::shared_ptr<Texturec> color) {
       m_brightTexture = color;
     }
-    
+
     inline std::shared_ptr<Texturec> darkTexture() const {
       return m_darkTexture;
     }
-    
+
     inline void setDarkTexture(std::shared_ptr<Texturec> color) {
       m_darkTexture = color;
     }
-    
+
+    /**
+      * Returns the coordinate mapping used to choose the bright or dark
+      * checker field.
+      */
+    inline const TextureMapping2D* mapping() const {
+      return m_mapping;
+    }
+
     virtual Colord evaluate(const Rayd& ray, const HitPoint& hitPoint) const;
-  
+
   private:
     TextureMapping2D* m_mapping;
     std::shared_ptr<Texturec> m_brightTexture;