+++ test/unit/render/textures/CheckerBoardTextureTest.cpp
@@ -8,29 +8,27 @@
 
 namespace CheckerBoardTextureTest {
   using namespace render;
-using namespace render;
-using namespace render;
 
   TEST(CheckerBoardTexture, ShouldInitializeWithValues) {
+    auto mapping = new PlanarMapping2D;
     auto bright = std::make_shared<ConstantColorTexture>(Colord::white());
     auto dark = std::make_shared<ConstantColorTexture>(Colord::black());
-    CheckerBoardTexture texture(new PlanarMapping2D, bright, dark);
+    CheckerBoardTexture texture(mapping, bright, dark);
+    ASSERT_EQ(mapping, texture.mapping());
     ASSERT_EQ(bright, texture.brightTexture());
     ASSERT_EQ(dark, texture.darkTexture());
   }
-  
+
   TEST(CheckerBoardTexture, ShouldChooseSubtextureBasedOnPosition) {
-    CheckerBoardTexture texture(
-      new PlanarMapping2D,
-      std::make_shared<ConstantColorTexture>(Colord::white()),
-      std::make_shared<ConstantColorTexture>(Colord::black())
-    );
-    
-    ASSERT_EQ(Colord::white(), texture.evaluate(
-      Rayd::undefined, HitPoint(nullptr, 0, Vector4d(0.5, 0, 0.5), Vector3d::null))
-    );
-    ASSERT_EQ(Colord::black(), texture.evaluate(
-      Rayd::undefined, HitPoint(nullptr, 0, Vector4d(1.5, 0, 0.5), Vector3d::null))
-    );
+    CheckerBoardTexture texture(new PlanarMapping2D,
+                                std::make_shared<ConstantColorTexture>(Colord::white()),
+                                std::make_shared<ConstantColorTexture>(Colord::black()));
+
+    ASSERT_EQ(Colord::white(),
+              texture.evaluate(Rayd::undefined,
+                               HitPoint(nullptr, 0, Vector4d(0.5, 0, 0.5), Vector3d::null)));
+    ASSERT_EQ(Colord::black(),
+              texture.evaluate(Rayd::undefined,
+                               HitPoint(nullptr, 0, Vector4d(1.5, 0, 0.5), Vector3d::null)));
   }
 }